Change manila DB to have working unique constraint
Because Manila DB uses 'soft delete', we can't use
unique constaints properly. Solution is to change
'deleted' column from Boolean to the type of 'id'
column and write value of 'id' to 'delete'
on delete operation. To apply unique constr only
to not deleted records we will be adding
'deleted' column to every uc from now on.

Closes-Bug: #1272303
Change-Id: Ia673b91c7d7f700a25824766486a82f07203f3b6