devstack: strip obsolete part of m-shr instumentation
The standard way of running services in screen windows
is the

$ cd $<SERVICE>_DIR && <SERVICE_CMD> --config-file $<SERVICE>_CONF

In case of the m-shr window (manila-share service) this was prefixed
with some initialization command to work around some initialization
context issues.

That's not required anymore, as those were specific to the lvm
driver, and now Manila is configured with lvm as default.
(If someone wishes to use lvm s/he can be referred to the cinder-volume
service which exhibits an issue analogous to that of Manila's lvm).

So we can go back to the standard format.

Change-Id: Ie7b3f08d5efcd8571d108b99ed766f10771398a2