Enable skipped tests from test_capacity_weigher.py
Some tests from  test_capacity_weigher.py were skipped:

- test_capacity_weight_multiplier1
- test_capacity_weight_multiplier2
- test_default_of_spreading_first

because of assertion error when comparing weight of the winner host
with expected value.

Fix this assertions taking into account that weihts of the hosts
are normalised during weighing, so the weight of the winner host
will be 1.0 * weight_multiplier in case of positive weight_multiplier
and 0 in other case.

Make CapacityWeigher able to redefine weight_multiplier from
BaseWeigher.

Closes-Bug: #1329718

Change-Id: I504ae7d980aa063ffddb7f2676e91da4f790e0e4