Adds undocumented policies and defaults in policy.json
Some policies for specific actions exist in Manila but are not documented in
etc/manila/policy.json. This makes someone have to dig to see what is available
when looking quickly.

The fix adds all actions from the APIs to policy.json, with default policy for
most of the newly added actions, and rule:admin_api for the rest. The fix
changes the share_network and security_service actions to default. The fix also
removes four actions no longer in the API. The actions quotas:update_for_user
and quotas:update_for_project are replaced with quotas:update. The actions
share_network:activate and share:network:deactivate have been removed.

Change-Id: I2e9510d59d0d3b70974f1a851ee5d2246154caf4
Closes-Bug: #1326797