Setup for translation
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports of po files will be done by the OpenStack Proposal
bot once manila has moved to openstack and setup the usual
translation jobs as part of gating. This change ensures that the gating
jobs will do the right thing.

Change-Id: I71ca9e8c0e4ea5586c1f1616607dddec595f65b6