Remove manila-clear-rabbit-queues
This script should be removed because it does not
appear to be widely used and there is no user documentation for
it. Also, RabbitMQ now includes a management plugin and CLI,
rabbitmqadmin, that allows you to delete queues and exchanges.

For reference, this removed in nova as part of
https://review.openstack.org/#/c/40355/.

and removed from cinder:
https://review.openstack.org/#/c/109786/

Change-Id: I32403890e190969657ae813e7d3e34a951030321