Fix update of backend details in cmode driver
There are dependency on backend details, that are stored in database,
in cmode driver. And if share server could not be created for some reason,
backend details won't be present in database. This fix updates database
at start of server creation, before sending request for creation.
There are still present possibility to get empty details, but it is
not related to driver itself, it is related to gap between creation of server
in database and call of driver's method for server creation. That could be,
for example, in case when share-network provided with wrong network data.

also added unittests to changed method due to absense of it.

Change-Id: Id1f8a4607b0dd09eb79b56b600088e8a70fbd47a
Closes-Bug: #1357273