Use oslosphinx and remove local copy of doc theme
A packaging review in Fedora found that we had a copy of jQuery code.
Apparently, other OpenStack projects applied a patch to remove this
code, but Manila forked Cinder code from before. The across-the-board
change in OpenStack missed Manila, because we weren't incubated yet.

This patch is almost an exact copy of Cinder patch, but not cherry-
picked because of additional changes, such as the move from oslo.sphinx
to oslosphinx.

Patch is verified to produce readable documentation.

Change-Id: Id0b57f89ae6b63b26b7f0c3fa9af5a529e92ac69