glusterfs: edit config option specifying volume
glusterfs driver has an option to specify the GlusterFS volume to be
mounted on the Manila host. The GlusterFS volume to be mounted is
presently specified in a file, whose path is the config option
glusterfs_volumes_config. Instead have a string valued config option
in manila.conf that directly takes the same input.

Change-Id: I3f8734174db71ef232620c55d4ecdd2b5d0192ad
Closes-Bug: #1365941