Fix tempest test with share server listing with no filters
We have test in tempest plugin that gets list of share servers
and verifies server being used is present, but also verifies
that all share servers have some specific states. It is incorrect
because share servers from whole cluster are taken. And these 'all'
share servers are not related to test.

Change-Id: Iae849cb64ca36448eb5fa28a2075de614fea475c