Allow deleting share with invalid share server in generic driver
In generic driver, if we do not specify the "--share-network" while
creating a new share(not from snapshot), the share will get error state
and it cannot be deleted(will be in "error_deleting" status).

Remove decorator "ensure_server" for method "delete_share" that is
dedicated mostly for share creation and allow delete share with invalid
share-server in Generic driver.

Change-Id: Id4f2138507a1b6c6c303837b6bc09c39c86559df
Closes-Bug: #1405362