Fix searching mechanism of share-networks within tempest
Now, if share-networks are required, Tempest searches for them by net and
subnet IDs. But if we create some additional share-network in concurrent test
suite it can be chosen by some another test suite. And if it is deleted we
get ShareNetworkNotFound error.

So, make Tempest distinguish share-networks by name.

Closes-Bug: #1406655

Change-Id: Iff82786954d46841c01456e7aebd4c67ef24dfad