Add factory for NetApp drivers
As we have done in Cinder, NetApp is planning additional Manila
drivers and would like to have a factory to simplify the user
experience of selecting a driver in manila.conf. The
implementation will be very similar to the Cinder one.  Included
is the Cinder map in the share manager that allows users to
change driver paths/names while maintaining backward
compatibility.

Implements blueprint netapp-driver-factory

Change-Id: I7dd401688e1bcccdde09b3ecf701924769ad2fe8