Add tracing facility to NetApp cDOT driver
It is useful for engineering to be able to see the API traffic
to and from the storage controllers, as well as to see the
progression of method calls through the driver. This work will
add a simple config option to enable various debug tracing
types to be written to the debug log at runtime.

Implements blueprint netapp-driver-tracing
Change-Id: Idf47a901abd434f256b71b2da454558878a769a2