Add support for HDFS native protocol driver
Add basic support for hdfs native protocol exports.

A manila share in this driver is a sub-dir in hdfs root dir.
Instances talk directly to the hdfs backend with 'hdfs' protocol.
Access to the share is allowed by user based access type, which
is aligned with HDFS ACLs to support access control of multiple
users and groups.

Implements blueprint hdfs-driver

Change-Id: Id9bd391d829e11d332b03c2245691ea42fc30bcc