Do not instantiate network plugin when not used by driver
We have two types of share drivers. First type supports creation of
share servers and second does not.
Then, we have approach of "network plugin" for handling of networking
and it is useful only with first type of share drivers.

So, we should restrict instantiation of network plugins with second type of
share drivers.
It will allow us to remove unused dependencies for running some share drivers.

Change-Id: Ib7d6a9ff1b6ce0d5d2bca350040363ea1461c419
Closes-Bug: #1417087