Generic Driver image supported protocols
Currently the generic driver is able to work with any images supplied,
but it is coded to use always NFS and CIFS helpers in order to interact
with the image. If a different image is used, for the purpose of other
share protocols, the above-mentioned helpers may not work. Since
image selection is up to ADM to set up in manila.conf, the config option
"share_helpers" should be used to determine which helpers should be
initialized. Specifying an unsupported combination (helper and image)
will now print a proper error message during initialization and prevent
any additional resources from being used.

Change-Id: If0bebecb5005e4225f9aebc0a8332834d8600a83
Closes-bug: #1406759