Manila cDOT netapp:thin_provisioned qualified extra spec
Manila has "share types", analogous to Cinder "volume types", that
may contain qualified extra-specs - extra-specs that have significance
only for the backend driver and which are not reported back to the
scheduler.

This commit implements the first of several qualified extra-specs:
  * netapp:thin_provisioned
This extra spec determines whether the backing NetApp flexvol for
a Manila share will be "thick" or "thin" provisioned, i.e. with or
without space guarantees.

This is a boolean extra spec, 'netapp:thin_provisioned=true' or
'netapp:thin_provisioned=false'.  If this qualified extra spec
is not supplied, then the default action is taken for the backing
flexvol, namely to provision it as "thick".

Partially implements blueprint cdot-qualified-specs

Change-Id: I9096ead88ce94afc84a0c8bc8315f07368241ec2