Replace TEMPEST_CONCURRENCY with Manila-specific var
Currently TEMPEST_CONCURRENCY is hardcoded to 12. This change
uses a different variable name to control concurrency, called
MANILA_TEMPEST_CONCURRENCY. The value still defaults to 12, but
can be overridden by the environment by 3rd party CI systems. A
different variable name is used to ensure that in upstream CI
it will not conflict with any existing variables.

Change-Id: Ia7c76a3ef18a6d114336417bfda08321f8d8971d