Manila Scheduler should read full driver capabilities
The scheduler should see all the data reported by a driver in
update_share_stats(), but there were a couple of fields from the
top-level capabilities dictionary that weren't being copied into
the pools.  In particular, the value driver_handles_share_servers
must be present in each pool, since that value is required in each
share type and all scheduling will fail if the value isn't present
on each pool.

Change-Id: Id388dceecfaaca3145fda82dd98386a7ee51adb9
Closes-Bug: #1430434