Set proper attr  "deleted" for ShareTypes model
The sqlalchemy model "ShareTypes" does not redefine the default model
attribute "deleted". The Default model defines it as an integer, but it
is expected to be a character field, as it is set by the DB migration
scripts. This leads to unhandled DB error when one attempts to delete
share type using PostgreSQL.

Change-Id: I791312d2ba25b678d920623a587c59d4b7c2e284
Closes-Bug: #1431335