Fix setting of extra specs for share types
Share types have one required extra spec - "driver_handles_share_servers".
And API disallows us to set some additional extra specs if we do not provide
required one too. It should not be so, because if we do not want to update
already existing spec then we should not be obligated to provide it.

Change-Id: Id2b2c9dfae3da551ae79e66199a84773f333b883
Closes-Bug: #1435819