NetApp cDOT driver is too strict in delete workflows
If a share or share server is not created successfully,
sometimes Manila will not save the vserver info in the
share or share server object. In such cases, when asked
to delete such a share or share server, the cDOT driver
should not protest about the lack of vserver info but
instead should log a warning and not raise an exception
(which leaves the object in an error_deleting state).

This patch addresses a number of issues in the delete,
share-server-delete, and snapshot-delete workflows where
the cDOT driver could unnecessarily raise an exception
when it should merely do nothing and allow the workflow
to proceed.

Change-Id: I54cf96b8a24ac5272b37bce2f5118551504a1699
Closes-Bug: #1438893