NetApp cDOT driver fails Tempest cleanup on clone workflows
During the create-from-snapshot workflow, the cDOT driver clones
the flexvol backing a share, and a job is started to split the
clone from its parent.  The Tempest tests that exercise this
feature attempt to delete all artifacts created for the test
(child flexvol, snapshot, parent flexvol) immediately after
the test assertions.  This rapid deletion wouldn't normally
happen in the field, but during the Tempest runs the clone
split is still occurring during the cleanup phase, so the
cleanup fails.  There isn't a great way to detect and prevent
the issue, so the workaround is to intercept the error and
retry a few times.

Closes-Bug: #1447339
Change-Id: If272eb3c8060c95510226672a8d1928bea0cc6cf