Remove direct DB calls from glusterfs_native driver
- Remove direct DB calls from methods create_snapshot() and
delete_snapshot().
- Move code from _update_gluster_vols_dict() method to
ensure_share() method.

Note that the original of this change was submitted as
a fix to #1444914; however, we've backported it to
Kilo for the purpose of fixing an other bug, as indicated
below.

Change-Id: I6c95f6d9361093d832a536971a460c3cdda44dcb
Closes-Bug: #1447692
(cherry picked from commit 4eb57f53c1c705fa3abf2ab0c2973fe709dbe6c8)