Fix 'AllocType' read failure in Huawei driver
The 'AllocType' have two values: Thin and Thick.
The following piece of code in the Huawei driver has
incorrect path and as a result the 'AllocType' is always
'None'. The correct values should be "Thin" or "Thick"
only.

'fstype = root.findtext('FILESYSTEM/AllocType')'

In code line 366: '"ALLOCTYPE": 1'
The "Thin" is set as the initial value. If the 'AllocType'
is always 'None', the 'AllocType' can not set as "Thick".

Change 'FILESYSTEM/AllocType' to 'Filesystem/AllocType' in
Huawei driver to avoid 'AllocType' can not be read out.

Change-Id: I63b77b8608b313729159d5732a5dae00cbe67034
Closes-Bug: #1451386
(cherry picked from commit c2a1fe786859314132a4fdfeb8ab3cb6e7ebdb46)