Fix tempest ShareUserRules* tests
Fix handling of return value from create_share 
in resource_setup.

These test classes (part of 'test_rules') are 
not currently used in CI and are not runnable 
because the resource_setup is still expecting a
tuple to be returned from create_share, but that 
method no longer returns a tuple.

This fix makes the following tests runnable:
- test_create_delete_user_rule
- test_create_delete_ro_access_rule

Change-Id: Ia4324e2bc9ea4cf80ed61e7c23110e9f4068a59b