Add Windows SMB share driver
This patch proposes a new share driver, handling Windows service
instances and exporting SMB shares.

WinRM is used for managing the service instances, allowing
password or certificate based authentication.

A new service instance manager is introduced, handling Windows
instances. One requirement is that Cloudbase-Init is present on the
instance.

If AD credentials are provided via security services, the instance
is joined to an AD domain. This can be leveraged in order to set
user based permissions.

A CI testing this driver will be up by L-3. Meanwhile, a link to
Tempest results can be found on the blueprint page whiteboard.

Change-Id: Ice5ee63ee4dc2cfcb41567214c165c2705421920
Implements: blueprint windows-smb-support