Add manage/unmanage support to NetApp cDOT driver
This commit adds support to the cDOT driver for the driver methods
manage_existing and unmanage.  The logic follows three sections:

1. Locate the resource to be managed, which is specified as the
cDOT flexvol name in the export_path argument.
2. Ensure the flexvol to be brought under Manila management is
of the appropriate type, the extra specs from the share type are
valid and match the specified pool, and the flexvol has no LUNs
or other flexvols mounted underneath it.
3. Rename the flexvol using normal Manila naming conventions,
remount it on a new path matching the new name, update it to
conform to all relevant qualified extra specs, and remove any
existing export rules.

Implements blueprint netapp-cdot-driver-manage-unmanage-share

Change-Id: Ib6832baf24585791bf7ec292e5c6cdeb83fe280d