Catch error_deleting state for more resources than just shares
The shares client in tempest currently catches if a share is in 'error_deleting'
state and throws a ResourceReleaseFailed exception so the tests don't have
to wait for the timeout.

However, more resources than just shares can go into error_deleting state,
snapshots for example.  This patch changes the check for error_deleting to
work with all resource types.

closes-bug: 1477703

Change-Id: Idc530db6593769df8f503878ce7432270ddae086