Use all types of migrations in devstack installation
For the moment we install Manila database using only 'upgrade' migrations.
In that case we do not test Manila after using 'downgrade' migrations.
So, allow enabling of both types of migrations to verify that our downgrade
migrations do not break the world of data having functional tests running
after it.

Change-Id: If55a38692abca5e008095eb99b55adb465cad2b6