glusterfs_native: prefix GlusterFS snap names with "manila-"
Due to change I40873208c7431e42885bee4db06d6229a202bad6,
we are free to change snapshot naming on the GlusterFS
side, as long as the Manila snap id occurs in it.

This is part of an effort to provide better
indication to the GlusterFS admin about the
origin of Manila created entities.

Partially implements bp manila-prefix-for-gluster-entities

Change-Id: Ie15e89cd49ab8450921ab08c78eb382096b57266