Update NetApp cDOT Manila drivers to use netapp_lib
Like other vendors, NetApp has published a storage controller
client library to PyPI for use by its OpenStack drivers.  This
patch updates the NetApp cDOT Manila drivers to use the new
library.  In particular, the unit tests must still work in the
gate in the absence of the new library.

Implements bp: use-pypi-client-library-in-netapp-cdot-drivers
Change-Id: I362f7ea584bb15a39c761fb5114391f557506764