Fix Generic driver share extend
When extending shares, the Generic share driver will extend
the underlying Cinder volume. It will consider this operation to
have finished once the volume status is "available".

In some cases, the volume status can be reported as active before
actually starting the extend operation. For this reason, attaching
the volume can fail as the volume would be in 'extending' state.

This patch fixes the issue by properly waiting for the volume
extend operation to finish, checking the volume size against the
expected size.

Change-Id: I02e4a42110e323dc295e20f9f07411a13fdcc236
Closes-Bug: #1479697