Consistency groups in NetApp cDOT drivers
With the consistency groups feature in Manila core, we can support
that in the NetApp cDOT drivers.  Note that cDOT does not have a
persistent CG concept, so some of the driver methods don't have much
to do; the backend merely takes the synchronized snapshots across
multiple shares as needed.

DocImpact
Implements bp: consistency-groups-in-netapp-cdot-drivers

Depends-On: Id6e56426b212435366a36122dbeddc52910d9828
Change-Id: I27b707ea8ce34d899d300c9a2b207738effe069a