NetApp cDOT driver with vserver creds can't create shares
With the new oversubscription-aware scheduler logic in Liberty,
the NetApp cDOT driver can no longer provision shares when the
driver only has vserver-scoped credentials.  This is because the
new scheduler logic requires the pool total size, whereas
previously the scheduler only needed the pool free space. Total
pool size is not available from cDOT when using vserver-scoped
credentials, so the driver must change to return 'unknown' in
that case and none of the free space checking or oversubscription
logic can be used.  When using cluster-scoped credentials, total
pool size is available and everything works fine.

Change-Id: I772fd8b133364439381a656ed84f9319b38eb135
Closes-Bug: 1490357