Capacity filter should check free space if total space is unknown
The Manila capacity filter accepts values of 'unknown' for both
total and free space in a pool. If free space is unknown, then
the scheduler always allows a create (as it should). But the
scheduler also always allows a create if total space is unknown,
even if free space is known; in this case, the filter should
compare the request against the available free space before
allowing the create.

Change-Id: Iad357f2472da1db493df42484c90cacf688608f8
Closes-Bug: 1490364