Fix HDS HNAS manage incorrect share size
Method get_quota does not handle quotas smaller than 1 GB.
When the share quota is KB or MB, get_quota transforms everything to GB.
It is impacting directly on managing shares.
As an example, managing a 500 MB share results in 500 GB share in HDS HNAS
Driver.
The unit TB was already being transformed in GB.

This patch implements a condition to deny share with quota smaller than
1GB.

Closes-bug: #1496917
Change-Id: I73edc1f413bd307a856702f8a024a8b50886207b