glusterfs_native: Hardwire Manila Host CN pattern
Currently, the glusterfs_native driver can support
create_share_from_snapshot API only if the Manila host's TLS
certificate's common name(CN) is prefixed by 'glusterfs-server'. This
constraint on the CN of the host is not acceptable. So fix this by
hardwiring a Manila host CN pattern that allows the Manila host CN
to have a more intuitive prefix, 'manila-host'.

Change-Id: Ibab27da4f9b28c6e7d2ef9c175d5decf5fa67ce4
Closes-Bug: #1496833