Can't create shares on drivers that don't support snapshots
For a driver that doesn't support snapshots, a share type with
snapshot_support = '<is> False' will prevent share creation.
Root cause is code in HostManager that improperly checks whether
snapshot_support (and other boolean properties) is set on each
pool. This commit fixes the check for the boolean pool capabilities.

Change-Id: I6a488788926cca119498b3de992beec3b0545259
Closes-Bug: #1498646