NetApp cDOT driver isn't reentrant
While testing the cDOT driver at high concurrency (12), we observed a few
occasional errors that were difficult to explain until we noticed that
some APIs were being sent to the wrong vserver. Root cause is that the
protocol helpers are cached and reused despite being stateful. There is
virtually no performance benefit to caching the helper classes, so the
simplest fix is to stop caching them.

Change-Id: I30efcf5a3ada4cc03d8506738876498ae75d20d3
Closes-Bug: #1500257