Use random IPs in security service tests
This patch changes the creation of security services in Manila
tempest to use randomized IP addresses instead of name-like strings
for both the 'dns_ip' and 'server' fields. The use of a name-like
string for the dns_ip field does not make sense from a positive test
perspective. The server field is more ambiguous in whether it should
take an IP or a hostname, but since nearly anything that can take a
hostname can also accept an IP, but not necessarily the reverse
(e.g. the NetApp drivers are expecting an IP), then it makes more
sense to use an IP here as well to cater to a broader audience.

The rand_ip() method is not in an ideal location. My preference
would be to push for this method to make it into the
tempest_lib.common.utils.data_utils package along with all the other
rand_whatever() methods. However, until that can be submitted,
accepted, and released, it will need to live inside the manila tests
temporarily.

Closes-bug: 1500899

Change-Id: I7dbeb9407e056266cc7a61722365f70ab668a525