Generic driver: wait for common server during setup
If the driver does not handle share servers, during init, it
verifies that the pre-configured share server exists and is available
(behavior recently introduced by
I325d6461c4b12f481769bb01f792ef850dba789c).

If the instance does not exist when the service starts, the driver
will remain permanently in an uninitialized state, even if the
according instance is created meanwhile.

This patch changes this behavior, waiting for the common server
to become active, logging a meaningful message. This can considerably
ease the deployment of this driver as well as other drivers based
on it.

At the moment, the Windows SMB driver CI jobs fail due to this issue,
when DHSS is set to false (in which case the Nova Hyper-V nodes are
deployed after Manila).

Change-Id: Icf446ca25cf0c64396223611c6271d4c6445cb4d
Closes-Bug: #1506951