HPE3PAR support for access-level (ro,rw)
NFS: 3PAR NFS shares can only be ro or rw, so create 2 shares
that both point to the same destination with the same export location.
This appears as one share in Manila. Create the RW share when the
share is first created (to make sure it is create-able) and add the
RO version when needed to support an RO access rule.

CIFS: For user access, this is easy. Just set read instead of fullcontrol.
Since 3PAR cannot do read-only at the CIFS IP level and also since most Manila
drivers do not use IP access on CIFS, take out the default requirement
of IP access for CIFS. Add a config setting to allow the old behavior to
be used if needed.

Implements: blueprint hpe3par-access-level

Change-Id: I284e53af777840d844ee9a32ac83355bb22595b9