Last sync to Manila from oslo-incubator
Oslo incubator is going away, and we cannot use the sync script
to copy the incubator files to Manila one last time because at
least one file has changed in Manila since the last sync.  This
commit includes the remaining changes which we do want.  The files
are consistent with the oslo-incubator master branch as of commit
305c2b0d52d387c0039e3b3021c4de27e85cd175.

Change-Id: Ib880a0108e4a567ebe39aa1ada7f872c63a818f1