Fix import of devstack functions for common CI script
Commit [1] moved common logic for CI hooks to separate file, but it
depends on some devstack functions, that should be "sourced" in this
common file. So, this common file works in manila hooks, because they
import devstack functions separately, but it will not work if some other
scripts use it and do not import devstack functions.

[1] I94514405628ab8337d1f4153e431f064614e3596

Change-Id: I919fea046e556d601a7680ea9bac16c1a9b7343c
Closes-Bug: #1519028