Check share-network in 'share create' API
For the moment it is possible to schedule share creation with DHSS=true
share type but without share network. But it makes no sense, and
expected to fail. So, perform check on API level.

APIImpact
When create share with share type(DHSS=true) and not input share
network, API will return HTTPBadRequest and message:
"Share network must be set when the driver_handles_share_servers
is true."

Closes-Bug: #1525125
Change-Id: Icdfabff7b1d3b6e95dd1dd58a0155de637056657