HPE3PAR finds CIFS share with either prefix
The share prefix can be osf- or osf-ro depending
on whether it is a readonly share. CIFS does not
create dual shares like NFS, but can use either
prefix because of the readonly snapshot special
case. So the CIFS implementation needs to find the
share or the "other share" for access rules.
Most of the NFS code is reused for this.

Unlike NFS a regular CIFS share supports both RO
and RW users.

An CIFS osf-ro share should only allow RO users.
In this case, find the share and later validate
the access rule.

Change-Id: I1e7a504c276e0e28f56a564802af17061fe09037
Closes-Bug: #1528936