Huawei driver report pool capabilities [True, False]
In liberty, huawei driver has been changed to support only
thin or thick filesystem in one storagepool because of bug#1491272.
Now in mitaka, after blueprint support-lists-in-capabilities-filter
has been merged. The scheduler is able to handle one storage pool
report "thin_provisioning" :[True, False].
Most of the code is the revert of bug#1491272.

Change-Id: Ibde3053be02274956e9a689845d5f3ef6c005c51
Closes-Bug: #1532651