Return appropriate data on share create
With the introduction of share instances in the Liberty
release, some attributes of the share model were moved
to the share_instance model. The create API has since
been returning merely the share data which makes
for a bad user experience with some attributes
being set to null.

Fix this issue by retrieving the share from the database
after the instance has been created; so that the status
field reflects the pre-scheduler acceptable status of
'creating'.

Change-Id: I63955e1a60e7ced3ad126829521cd43e0b914c17
Closes-Bug: #1534161