glusterManager instantiation regexp validation
GlusterManager can be instantiated in two ways:
- with an address string like "johndoe@example.com:/vol/dir"
- with a component dict like
   {'user': 'johndoe',
    'host': 'example.com',
    'volume':'vol',
    'path': 'dir'}

If the instance is created with an address string, it will be parsed
using regexp which facilitates the validation of the certain component
values.

Made changes such that if the GlusterManager is created using a
component dict, then change the format into the address string and allow
it to be parsed.

Mangling it through the parse ensures it meets the same criteria as
string addresses, without a partial reimplementation of parsing.

Closes-Bug: #1496733
Change-Id:I9128ab086d63b72f33a6e879cb038a6759b56ef9